+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.2] 2025-03-19
+
+* Support for flag `parkingSpotZoneId` of `includeAttributes` on `/v2/access-rights`.
+* Support for attribute `zoneId` when listing Parking Spots.
+* Upgrade Jackson from `2.17.2` to `2.18.3`.
+
 ## [2.7.1] 2024-09-13
 
 * Support for attribute `identificationValue` when creating Access Logs.